Move encode and decode function to another file
The encode and decode in the JS file seems to be for something else
and not for the content.  I still haven't figured out how to decode
the encoded content, the only sane way seems to be just rendering the
page using a JS-enabled library and going from there.  Reversing this
thing is a huge headache and beyond my skills.  The fact that all the
JS files are obfuscated does not help the situation.

* pratilipi2epub.py (decode_content):
(encode): Move from here...
* _encode_decode.py (decode_content, encode): to here.